backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
pillow
boto3
django-storages
